+++ chrommentor/src/tagsOptions.html
@@ -18,6 +18,11 @@
     <div>
       <button id="export">Export tags</button>
     </div>
+    <div>
+      <label for="import">Import tags:</label>
+
+      <input type="file" id="import" name="import" accept="application/json" />
+    </div>
     <table id="tags-options-table">
       <tr>
         <th>Label</th>
@@ -29,4 +34,5 @@
   </body>
   <script src="tagsOptions.js"></script>
   <script src="exportTags.js"></script>
+  <script src="importTags.js"></script>
 </html>
